+++ packages/sit-onyx/src/components/OnyxAccordion/OnyxAccordion.stories.ts
@@ -11,7 +11,6 @@ import OnyxAccordion from "./OnyxAccordion.vue";
 const meta: Meta<typeof OnyxAccordion> = {
   title: "Basic/Accordion",
   component: OnyxAccordion,
-  tags: ["new:component"],
   argTypes: {
     default: {
       control: { disable: true },